chat_with_doc_api/requirements.txt
gradio
mdtex2html
pypinyin
tiktoken
socksio
tqdm
colorama
accelerate
langchain
torch
langchain-serve
protobuf
faiss-cpu
sentence_transformers 
InstructorEmbedding
python-dotenv
openai
gpt4all
pyllama 
git+https://github.com/huggingface/peft.git
git+https://github.com/huggingface/transformers.git
SentencePiece
isort
black
pygpt4all
tiktoken
safetensors
xformers
bitsandbytes
einops
gevent
pydantic >= 1.10.11
pypdf
python-telegram-bot
transformers_stream_generator
openllm
openllm[llama]
text_generation
